skills component updated
