+++ src/components/sections/skills/skills.module.css
@@ -1,0 +1,273 @@
+.root {
+  background-color: var(--whitebg);
+  width: 100%;
+  border-bottom-left-radius: 40px;
+  border-bottom-right-radius: 40px;
+  padding-top: 2.5rem;
+  padding-bottom: 5rem;
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #1a1a1a;
+  }
+
+  @media (min-width: 768px) {
+    padding-top: 3.5rem;
+    padding-bottom: 6rem;
+  }
+
+  @media (min-width: 1024px) {
+    border-bottom-left-radius: 50px;
+    border-bottom-right-radius: 50px;
+    padding-bottom: 7rem;
+  }
+
+  @media (min-width: 1280px) {
+    padding-top: 5rem;
+    padding-bottom: 8rem;
+  }
+
+  @media (min-width: 1536px) {
+    border-bottom-left-radius: 70px;
+    border-bottom-right-radius: 70px;
+  }
+}
+
+.container {
+  width: 100%;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 1920px;
+  box-sizing: border-box;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+
+  @media (min-width: 768px) {
+    gap: 3.5rem;
+  }
+
+  @media (min-width: 1536px) {
+    gap: 5rem;
+  }
+}
+
+.title {
+  font-family: var(--font-heading);
+  text-align: center;
+  font-size: 1.5rem;
+  font-weight: 600;
+
+  @media (min-width: 640px) {
+    font-size: 1.875rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 2.25rem;
+  }
+
+  @media (min-width: 1280px) {
+    font-size: 3rem;
+  }
+}
+
+.carousel {
+  position: relative;
+
+  @media (min-width: 1280px) {
+    margin-left: auto;
+    margin-right: auto;
+    width: 83.333333%;
+  }
+}
+
+.fade {
+  pointer-events: none;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  z-index: 10;
+  width: 6rem;
+
+  @media (min-width: 768px) {
+    width: 6rem;
+  }
+
+  @media (min-width: 1024px) {
+    width: 7rem;
+  }
+
+  @media (min-width: 1536px) {
+    width: 8rem;
+  }
+}
+
+.left {
+  composes: fade;
+  left: 0;
+  background-image: linear-gradient(to right, var(--whitebg), transparent);
+
+  @media (prefers-color-scheme: dark) {
+    background-image: linear-gradient(to right, #1a1a1a, transparent);
+  }
+}
+
+.right {
+  composes: fade;
+  right: 0;
+  background-image: linear-gradient(to left, var(--whitebg), transparent);
+
+  @media (prefers-color-scheme: dark) {
+    background-image: linear-gradient(to left, #1a1a1a, transparent);
+  }
+}
+
+.card {
+  position: relative;
+  display: flex;
+  aspect-ratio: 1 / 1;
+  align-items: center;
+  justify-content: center;
+  border-radius: 0.5rem;
+  background-color: #ebebeb;
+  padding: 1rem;
+  transition: all 300ms;
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #2c2c2c;
+  }
+
+  @media (min-width: 640px) {
+    padding: 1.5rem;
+  }
+
+  &:hover {
+    transform: scale(0.95);
+  }
+
+  &:hover .logo {
+    transform: translateY(-0.5rem) scale(0.9);
+    opacity: 0;
+  }
+
+  &:hover .label {
+    bottom: 0.5rem;
+    opacity: 1;
+  }
+
+  &:hover .link,
+  &:hover .badge {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+.logo {
+  height: 66.666667%;
+  width: 66.666667%;
+  transition: all 300ms;
+
+  @media (min-width: 640px) {
+    height: 3rem;
+    width: 3rem;
+  }
+
+  @media (min-width: 768px) {
+    height: 3.5rem;
+    width: 3.5rem;
+  }
+
+  @media (min-width: 1536px) {
+    height: 4rem;
+    width: 4rem;
+  }
+}
+
+.label {
+  font-family: var(--font-heading);
+  color: color-mix(in srgb, var(--foreground) 80%, transparent);
+  position: absolute;
+  bottom: 0.25rem;
+  font-size: 8px;
+  font-weight: 500;
+  opacity: 0;
+  transition: all 300ms;
+
+  @media (min-width: 640px) {
+    bottom: 0.25rem;
+    font-size: 0.75rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 0.875rem;
+  }
+}
+
+.link-root,
+.badge-root {
+  font-family: var(--font-heading);
+  position: absolute;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 9999px;
+  background-color: #f5f5f5;
+  color: #000;
+  opacity: 0;
+  box-shadow:
+    0 10px 15px -3px rgba(0, 0, 0, 0.1),
+    0 4px 6px -2px rgba(0, 0, 0, 0.05);
+  transition: all 500ms;
+  transform: translateY(1rem);
+  text-decoration: none;
+  font-weight: 700;
+  padding: 0.25rem 0.75rem;
+
+  @media (min-width: 768px) {
+    padding: 0.5rem 1rem;
+  }
+}
+
+.link {
+  composes: link-root;
+  font-size: 8px;
+
+  @media (min-width: 460px) {
+    font-size: 10px;
+  }
+
+  @media (min-width: 640px) {
+    font-size: 0.75rem;
+  }
+
+  @media (min-width: 1024px) {
+    font-size: 0.875rem;
+  }
+}
+
+.badge {
+  composes: badge-root;
+  font-size: 0.75rem;
+
+  @media (min-width: 1024px) {
+    font-size: 0.875rem;
+  }
+}
+
+.icon {
+  height: 0.5rem;
+  width: 0.5rem;
+  margin-left: 0.25rem;
+
+  @media (min-width: 460px) {
+    height: 0.75rem;
+    width: 0.75rem;
+  }
+
+  @media (min-width: 768px) {
+    height: 1rem;
+    width: 1rem;
+  }
+}
